requirements.txt
Flask==1.1.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
numpy>=1.18.1
opencv-python==4.2.0.32
tensorflow=2.1.0
Pillow=7.0.0

